pyproject.toml
[project]
name = "netcommander"
version = "2025.10.03"
description = "Home Assistant custom component for controlling Synaccess netCommander Power Distribution Units (PDUs)"
authors = [{name = "Robert Sigler", email = "code@sigler.io"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "pydantic>=1.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
cli = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]
ha = [
    "homeassistant>=2024.1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
netcommander = "netcommander_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "C90", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # HA doesn't enforce strict typing
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src/netcommander --cov=custom_components/netcommander --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
